renaming domains directory to apps
