+++ src/apps/landingPage/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import NavBar from '@Domain/landingPage/components/navBar';
+import NavBar from '@App/landingPage/components/navBar';
 
 export default function LandingPage() {
   return (
